requirements.txt
asttokens==2.0.8
autopep8==2.0.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
filelock==3.8.0
fonttools==4.37.4
gensim==4.2.0
huggingface-hub==0.10.1
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
jupyter-client==7.4.1
jupyter-core==4.11.1
kiwisolver==1.4.4
matplotlib==3.6.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nltk==3.7
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0
smart-open==6.2.0
stack-data==0.5.1
threadpoolctl==3.1.0
tokenizers==0.13.1
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.24.0
typing-extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.12
wcwidth==0.2.5
